Add 0.5fps and 1fps options to live waveform refresh rate

+++ app.js
@@ -1099,7 +1099,7 @@
   });
 
   liveFpsSelect.addEventListener("change", function () {
-    liveTargetFps = parseInt(liveFpsSelect.value) || 30;
+    liveTargetFps = parseFloat(liveFpsSelect.value) || 30;
     liveFrameInterval = 1000 / liveTargetFps;
   });
